requirements.txt
requests
python-dotenv
schedule
openai>=1.0.0
tweepy